backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-dotenv==1.0.1
pydantic==2.9.2
numpy==2.1.2
Pillow==10.4.0
opencv-python-headless==4.10.0.84
mss==9.0.1
pytesseract==0.3.13
websockets==12.0
httpx==0.27.2
openai==1.51.0
pytest==8.3.3
